Extract user menu
